Drop beads task tracking from the repo
Delete the bip tree command, the flow.Bead type and its helpers, the
HTML tree generator + tests, the /bip-tree skill, and the .gitattributes
beads merge driver. Strip beads mentions from CLAUDE.md and the docs
guides, and fix the stale `bip board sync` doc example (no such
subcommand).

beads.go also housed the non-beads helper parsePositiveInt (used by
duration.go); relocated it and its errInvalidInteger dependency into
duration.go. mustParseInt and gitHubRefPattern had no remaining
consumers and were dropped.

Closes #139.
