--- .gitattributes
@@ -1,3 +1,0 @@
-
-# Use bd merge for beads JSONL files
-.beads/issues.jsonl merge=beads
